MT#6209 Install mirror signing key inside piuparts chroot
Problem as noted in several piupart runs since 7th of March:

| After purging files have been modified: /usr/lib/perl/5.14.2/Config_heavy.pl owned by: perl-base

perl[-base] 5.14.2-21+deb7u1 was uploaded on 2013-09-29 (see
http://packages.qa.debian.org/p/perl/news/20130929T211706Z.html),
while our piuparts base.tgz files for wheezy have been created on
2013-08-19 with perl[-base] version 5.14.2-21. This means that on
each piuparts run the perl-base package needs to be updated.

When switching to our own Debian mirror using deb.sipwise.com
(see MT#6171) this resulted in the following warning message:

| W: A error occurred during the signature verification. The repository is not updated and the previous index files will be used. GPG error: http://deb.sip
| W: Failed to fetch http://deb.sipwise.com/debian/dists/wheezy/Release

This wasn't considered as a fatal error. But as a result our
self-signed Debian repository caused apt-get to not consider
packages from deb.sipwise.com for inclusion in the dist-upgrade
process which is executed by piuparts.

When installing the "Debian-package-under-test" though the
perl-base package was considered as part of the installation, but
perl-base was also upgraded during this installation process.
The perl-base package actually *has* this flaw within the
5.14.2-21->5.14.2-21+deb7u1 upgrade, but it's not our problem.