MT#9905 Make sure piuparts chroot is up2date
The tarball that is used for unpacking doesn't include Sipwise'
apt-key so (dist-)upgrading it fails and the wrong checksums
are used. E.g. our tarball ships an older version of perl-base
than you get with the 7.7 point release nowadays and therefore
we get:

| After purging files have been modified: /usr/share/doc/perl/changelog.Debian.gz owned by: perl-base

as a result. By ensuring that our chroot is properly updated
we can avoid this issue.

To use this feature it requires the patch I provided in
Debian's #767485, I'll provide an according backport on our
infrastructure.