MT#11099 Drop usage of foreign jessie mirror + fix debootstrap option handling
We've our own Debian/jessie mirror nowadays, so there's no
need to use an external mirror for that anymore.

Using "DEBOOTSTRAP_OPTIONS=${!debootstrap_options}" doesn't work
(anymore, for unknown reasons). Since we need wget for further
checks we want to include it in our default environment. Also
we need to pass our trusted.gpg keyring so that bootstrapping
the piuparts environment works in a trusted way.