MT#15067 Ensure that chroot is up2date also with --existing-chroot
Before piuparts is recording its metadata of the chroot it's
supposed to be brought up2date, so that installed packages match
what's provided by the repositories being refered to in
sources.list.

piuparts until and including v0.66 executes the apt-get
dist-upgrade only when either the basetgz or schroot option is
present. We've switched from the basetgz approach to
--existing-chroot= though, because that speeds up build times.

With the recent Debian point release the libsqlite3-0 package
(which is a dependency of aptitude and therefore pre-installed in
the chroot) received an update which caused our piuparts runs to
fail because of wrongly recorded metadata, as mentioned above.

Reported against upstream including a patch in http://bugs.debian.org/798266

Change-Id: I4427d339f163bd92756e6ea71f0c0c99b766be7f