MT#16989 ngcp-upgrade doesn't use release specific packages anymore
Nowadays it's just ngcp-upgrade-ce + ngcp-upgrade-pro, so the
workaround to just run the checks for the current latest +
latest LTS release isn't needed anymore.

FTR: if this should fail for older releases we have to investigate
how to support (or ignore) that, but for now let's make sure that
master builds work as needed.

Change-Id: I86a991e6dbcf60c614111e740dfef1af1f3abd4c